server/requirements.txt
bs4
Django>=3.2,<4.0
django-cors-headers
django-environ
django-rest-framework
djongo
epubcheck
flake8
isort==5.8.0
pycodestyle
pymongo==3.12.0
pytest
sqlparse==0.2.4
unittest2
